+++ crates/core/src/cat/mod.rs
@@ -1,0 +1,4 @@
+//! This module defines the SurrealDB Catalog.
+#![deny(missing_docs)]
+
+pub mod namespace;
