+++ crates/core/src/kvs/version/fixes/v2_to_v3_data_model_migration.rs
@@ -3,7 +3,7 @@ use std::sync::Arc;
 use crate::{
 	err::Error,
 	kvs::{util, Transaction},
-	mdl::namespace::Namespace,
+	cat::namespace::Namespace,
 	sql::statements::DefineNamespaceStatement,
 };
