--- crates/core/src/mdl/mod.rs
@@ -1,4 +1,0 @@
-//! This module defines the SurrealDB data model.
-#![deny(missing_docs)]
-
-pub mod namespace;
