+++ revision.lock
@@ -137,8 +137,8 @@ MTreeParams:2(crates/core/src/sql/index.rs)(1827809571)
 Method:1(crates/core/src/api/method.rs)(2419610802)
 Mock:1(crates/core/src/sql/mock.rs)(4202853420)
 Model:1(crates/core/src/sql/model.rs)(4053086513)
-Namespace:1(crates/core/src/mdl/namespace.rs)(3182423989)
-NamespaceId:1(crates/core/src/mdl/namespace.rs)(42975151)
+Namespace:1(crates/core/src/cat/namespace.rs)(3793716955)
+NamespaceId:1(crates/core/src/cat/namespace.rs)(42975151)
 Node:2(crates/core/src/dbs/node.rs)(433772837)
 Notification:1(crates/core/src/dbs/notification.rs)(2307939178)
 Number:1(crates/core/src/sql/number.rs)(246551521)
